pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tcm-cli"
version = "0.1.0"
description = "TCM CLI — An autonomous agent for Traditional Chinese Medicine research and discovery"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "TCM-CLI Contributors" },
]

dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "prompt-toolkit>=3.0",
    "anthropic>=0.40",
    "openai>=1.0",
    "httpx>=0.27",
    "pandas>=2.0",
    "numpy>=1.24",
    "python-dotenv>=1.0",
    "markdown>=3.5",
]

[project.optional-dependencies]
chemistry = [
    "rdkit>=2023.03",
]
ml = [
    "torch>=2.0",
    "transformers>=4.40",
]
analysis = [
    "seaborn>=0.13",
    "scikit-learn>=1.3",
    "scipy>=1.10",
]
all = [
    "rdkit>=2023.03",
    "torch>=2.0",
    "transformers>=4.40",
    "seaborn>=0.13",
    "scikit-learn>=1.3",
    "scipy>=1.10",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
tcm = "tcm.cli:entry"

[tool.hatch.build.targets.wheel]
packages = ["src/tcm"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "e2e: end-to-end tests hitting real APIs",
]
